add Automated Email
This script uses the mail command to send an email. The recipient's email address, subject line, and message body are defined as variables at the beginning of the script. The message is sent using the echo command, which pipes the message to the mail command. The -s option sets the subject line. Finally, a confirmation message is displayed to indicate that the email was sent.