add  script that creates a tar archive of a specified directory
This script uses the tar command to create a gzipped tar archive of the specified directory. The directory to backup and the backup file name are defined as variables at the beginning of the script. The backup file name includes the current date in its name to differentiate it from previous backups. The -z option compresses the archive using gzip, the -c option creates a new archive, and the -v option provides verbose output while the archive is being created. Finally, a confirmation message is displayed to indicate that the backup was created.