Add README for script concatenate all the video
This script uses the ffmpeg tool to concatenate all the video files in the directory specified by src_dir into a single video. The -f option specifies the format of the input files, and the -i option specifies the input files. The input files are specified using a shell command that generates a list of file directives for each video file in the directory. The -c option specifies the codec to be used for the output file, in this case copy to copy the video streams directly from the input files without re-encoding. The output video is saved to the file specified by output_file.