Add Command-Line Video Player
This script uses the mplayer command-line video player to play the video file specified by input_file. The -vo option specifies the video output driver to be used, in this case caca to display the video in the terminal using the caca-utils package.

Note that the caca-utils package must be installed on your system in order to run this script. You can install it using a package manager such as apt-get on Debian-based systems or yum on Red Hat-based systems.